+++ CHANGES.rst
@@ -4,7 +4,9 @@ Changelog
 1.1 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Add a subject indexer, which adds EMPTY_MARKER to the index, where no subject
+  is set. Make sure, the EMPTY_MARKER is translated in the portlet.
+  [thet]
 
 
 1.0 (2015-07-15)